Tests/LibWeb/Text/expected/css/media-query-serialization-basic.txt

@media screen {
}
@media screen and ((min-width: 20px) and (max-width: 40px)) {
}
@media screen and (min-resolution: 1dppx) {
}
@media screen and (min-resolution: 2dppx) {
}
@media screen and (min-resolution: 2.54dppx) {
}
